# util/swagger_2025_01_28_candidate.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Jcurve APIs",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://qaapi.jcurve.tech/api/v1",
      "description": "QA",
      "variables": {}
    },
    {
      "url": "https://api.jcurve.tech/api/v1",
      "description": "Production",
      "variables": {}
    }
  ],
  "paths": {
    "/validateEmail": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "send validation link to email",
        "operationId": "sendvalidationlinktoemail",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/sendvalidationlinktoemailrequest"
                  },
                  {
                    "example": {
                      "email": "arjunnair@yopmail.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "arjunnair@yopmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/verifyEmail/{token}": {
      "put": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "verify email",
        "operationId": "verifyemail",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "signup user",
        "operationId": "signupuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/signupuserrequest"
                  },
                  {
                    "example": {
                      "email": "arjunnair@yopmail.com",
                      "firstName": "Arjun",
                      "lastName": "nair(Dummy)",
                      "password": "Arjun@123"
                    }
                  }
                ]
              },
              "example": {
                "email": "arjunnair@yopmail.com",
                "firstName": "Arjun",
                "lastName": "nair(Dummy)",
                "password": "Arjun@123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/webLogin": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "login",
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginrequest"
                  },
                  {
                    "example": {
                      "email": "arjunnair@yopmail.com",
                      "password": "Arjun@123"
                    }
                  }
                ]
              },
              "example": {
                "email": "arjunnair@yopmail.com",
                "password": "Arjun@123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/forgotPassword": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "forgot password",
        "operationId": "forgotpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/forgotpasswordrequest"
                  },
                  {
                    "example": {
                      "email": "arjunnair@yopmail.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "arjunnair@yopmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/verifyForgotPasswordToken/{token}": {
      "put": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "verify forgot password token",
        "operationId": "verifyforgotpasswordtoken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/updatePassword/{userId}/{token}": {
      "put": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "update password",
        "operationId": "updatepassword",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/updatepasswordrequest"
                  },
                  {
                    "example": {
                      "password": ""
                    }
                  }
                ]
              },
              "example": {
                "password": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/changePassword": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "change password",
        "operationId": "changepassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/changepasswordrequest"
                  },
                  {
                    "example": {
                      "oldPassword": "",
                      "newPassword": ""
                    }
                  }
                ]
              },
              "example": {
                "oldPassword": "",
                "newPassword": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/logout": {
      "post": {
        "tags": [
          "Candidate Registration & Login"
        ],
        "summary": "Logout",
        "operationId": "Logout",
        "parameters": [
          {
            "name": "partnerCode",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "C001P01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/student/getJobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "getJobs",
        "operationId": "getJobs",
        "parameters": [
          {
            "name": "employmentType",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "Full Time"
            }
          },
          {
            "name": "partnerCode",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "C001P01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/companies": {
      "get": {
        "tags": [
          "Masters  (drop down apis)"
        ],
        "summary": "companies",
        "operationId": "companies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/collegesList": {
      "get": {
        "tags": [
          "Masters  (drop down apis)"
        ],
        "summary": "get colleges list",
        "operationId": "getcollegeslist",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "for searching college names",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "sendvalidationlinktoemailrequest": {
        "title": "sendvalidationlinktoemailrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "arjunnair@yopmail.com"
        }
      },
      "signupuserrequest": {
        "title": "signupuserrequest",
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "arjunnair@yopmail.com",
          "firstName": "Arjun",
          "lastName": "nair(Dummy)",
          "password": "Arjun@123"
        }
      },
      "loginrequest": {
        "title": "loginrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "arjunnair@yopmail.com",
          "password": "Arjun@123"
        }
      },
      "forgotpasswordrequest": {
        "title": "forgotpasswordrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "arjunnair@yopmail.com"
        }
      },
      "updatepasswordrequest": {
        "title": "updatepasswordrequest",
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "example": {
          "password": ""
        }
      },
      "changepasswordrequest": {
        "title": "changepasswordrequest",
        "required": [
          "oldPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "example": {
          "oldPassword": "",
          "newPassword": ""
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}